Confirm env variables. Add verify-constraint to tests. Add shell option to menu. Fix output dir.  Other fixes

+++ tools/release-kit/test_pp.sh
@@ -5,8 +5,7 @@ source "./utils.sh"
 source "$HOME/.bashrc"
 
 load_env
-info "Press any key to continue..."
-read -n 1 -s
+confirm_continue
 
 ELF_FILE="pessimistic-proof-program-keccakf-v0.7.0.elf"
 MPI_CMD="mpirun --bind-to none -np $DISTRIBUTED_PROCESSES -x OMP_NUM_THREADS=$DISTRIBUTED_THREADS"
@@ -28,22 +27,29 @@ if ! grep -F "ROM setup successfully completed" romsetup_output.log; then
     err "program setup failed"
     exit 1
 fi
-  
-step "Generating pessimistic proof (no distributed)..."  
+
+step "Verifying constraints for pp_input_1_1.bin..."
+ensure cargo-zisk verify-constraints -e "pessimistic-proof/program/${ELF_FILE}" -i "pessimistic-proof/inputs/bench/pp_input_1_1.bin" 2>&1 | tee constraints_output.log
+if ! grep -F "All global constraints were successfully verified" constraints_output.log; then
+    err "verify constraints failed"
+    exit 1
+fi
+
+step "Generating pessimistic proof for pp_input_1_1.bin (no distributed)..."  
 ensure cargo-zisk prove -e "pessimistic-proof/program/${ELF_FILE}" -i "pessimistic-proof/inputs/bench/pp_input_1_1.bin" -o proof -a -y 2>&1 | tee prove_output.log
 if ! grep -F "Vadcop Final proof was verified" prove_output.log; then
     err "prove program failed"
     exit 1
 fi
 
-step "Verifying pressimistic proof..."
+step "Verifying pressimistic proof for pp_input_1_1.bin..."
 ensure cargo-zisk verify -p ./proof/proofs/vadcop_final_proof.json -u ./proof/publics.json 2>&1 | tee verify_output.log
 if ! grep -F "Stark proof was verified" verify_output.log; then
     err "verify proof failed"
     exit 1
 fi
 
-step "Generating pessimistic proof (distributed)..."  
+step "Generating pessimistic proof for pp_input_20_20.bin (distributed)..."  
 ensure $MPI_CMD cargo-zisk prove -e "pessimistic-proof/program/${ELF_FILE}" -i "pessimistic-proof/inputs/bench/pp_input_20_20.bin" -o proof -a -y 2>&1 | tee prove_output.log
 if ! grep -F "Vadcop Final proof was verified" prove_output.log; then
     err "prove program failed"
